- files_utils.py
Updated to include a super generic set of user headers to just fool YouTube at a basic sense of course.

- reuploads.py

Simple update to include youtube support on the website command.

- reuploads_utils.py

Massive update here on the backend to make sure that support is there for youtube. It'll likely break at some point in the future, but for now it works on the relatively few channels I put it through.